Make a newly created game include its boards in the JSON response
This is a very ugly solution. Upon creating a game, we need to actually
create the game (which returns a promise), set its boards (which returns
another promise) and respond to the request when both of these promises
have finished. Furthermore, it seems that we need to

  include: [models.Board]

in order for a game's boards to be included in the response, which means
that we also need to do a `models.Game.find()`.

Because the board attributes are set asynchronously, the game which is
returned from `startGame()` does not necessarily (and, in fact, does
not) have its boards set at the time of return, resulting in an empty
array being returned for the `boards` attribute. The only workaround
I've found so far is to perform yet another `find()`, which runs after
the appropriate boards are in place. Whether this is a reliable
solution, or if it's subject to race conditions, I'm not actually sure.
It is, however, both ugly and currently working. So I'm pushing it,
hoping to find a better solution in the future.